Foundation 0.2
The first, half completed version of the "foundation" for the game. I am quite satisfied with it. It is quite awesome. Comes with a small demo of printing "Hello World" in different colors. Does not look as impressive as the older ones, but inside it works twice as impressive as the older ones.